Per-arch Split of the glibc_elf.h
